refactor(cron,loop): extract dep-free leaves into @openloomi/cron + @openloomi/loop
Phase 5 of the runtime/UI split. Two new zero-dep leaf packages own the
pieces of `apps/web/lib/cron/` and `apps/web/lib/loop/` that have no DB,
agent, or integration dependencies, so the runtime + UI can both import
the same primitives without dragging the rest of those folders along.

`@openloomi/cron` (3 files moved):
  - `types`              — `ScheduleConfig`, `JobConfig`, `CronJob`,
                            `SchedulerEvent`, `SchedulerConfig`, etc.
                            The original Drizzle `ScheduledJob` import
                            is replaced with a structural
                            `ScheduledJobLike` interface so the package
                            stays free of `apps/web/lib/db/schema.ts`.
  - `scheduler`          — `computeNextRun`, `validateCronExpression`,
                            `isJobDue`, `formatDate`, `parseDate` (croner).
  - `stream-response`    — `createJobExecutionStreamResponse` (SSE).

`@openloomi/loop` (3 files moved + 1 inlined type/constant):
  - `paths`              — `LOOP_HOME`, `LOOP_PATHS`, `ensureDirs`,
                            `migrate`, `ensureParent` (pure node:fs/path).
  - `cli-path`           — `resolveLoopCli`, `listLoopCliCandidates`
                            (resolves the `loop-cli.mjs` shim across
                             dev / standalone / packaged Tauri modes).
  - `preferences`        — `readPreferences`, `writePreferences`. The
                            `LoopPreferences` interface and
                            `DEFAULT_LOOP_PREFERENCES` constant move
                            here too so the package is self-contained.
                            `apps/web/lib/loop/types.ts` re-exports
                            them for back-compat with the rest of the
                            Loop runtime.

Local files in `apps/web/lib/{cron,loop}/` are now one-line re-export
shims. Path aliases added to `apps/web/{tsconfig,vitest.config}.ts` and
workspace deps added to `apps/web/package.json`.

The new `apps/web/lib/loop/preferences.ts` and `apps/web/lib/loop/types.ts`
now use `import type` alongside the re-export so later references in
the same file (e.g. `LoopState.preferences: LoopPreferences`) still
resolve.

Test fix: `tests/unit/loop/notifications.test.ts` previously mocked
`@/lib/loop/paths` to redirect file I/O to a tmp dir. The mock now also
covers `@openloomi/loop/paths` so the new package's `readPreferences`/
`writePreferences` see the same tmp dir. The other 16 test files that
mock `@/lib/loop/paths` don't touch `preferences.ts` so they keep
working unchanged.

Verification: `tsc --noEmit` clean, `prettier --check` clean on all
new and modified files, `pnpm test` 2253 passed / 17 skipped / 0 failed.
