+++ apps/web/package.json
@@ -51,11 +51,13 @@
 		"@openloomi/ai": "workspace:*",
 		"@openloomi/api": "workspace:*",
 		"@openloomi/audit": "workspace:*",
+		"@openloomi/cron": "workspace:*",
 		"@openloomi/hooks": "workspace:*",
 		"@openloomi/i18n": "workspace:*",
 		"@openloomi/indexeddb": "workspace:*",
 		"@openloomi/insights": "workspace:*",
 		"@openloomi/integrations": "workspace:*",
+		"@openloomi/loop": "workspace:*",
 		"@openloomi/memory-consolidation": "workspace:*",
 		"@openloomi/memory-store": "workspace:*",
 		"@openloomi/rag": "workspace:*",
