+++ packages/cron/src/index.ts
@@ -1,0 +1,6 @@
+// @openloomi/cron — scheduling primitives shared by runtime + UI
+// See `docs/split-runtime-ui.md` for boundary contract.
+
+export * from "./types";
+export * from "./scheduler";
+export * from "./stream-response";
